Can remove entries from added food

+++ Model/UserInfoModel.swift
@@ -43,9 +43,9 @@ class UserInfoModel: ObservableObject {
     
     struct CurrentCalorieProgress: Identifiable{
         var id = UUID()
-        var calorieProgress: Int
-        var fatProgress: Int
-        var carbProgress: Int
+        var calorieProgress: Double
+        var fatProgress: Double
+        var carbProgress: Double
     }
     
     @Published var personUserInfo = UserInfo.init(firstName: "",  height: 0, weight: 0, gender: "", age: 0, activityLevel: "", BMR: 0)
